requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.1
coverage==7.13.1
Django==5.0
django-prometheus==2.4.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
exceptiongroup==1.3.1
fakeredis==2.33.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.0
gunicorn==21.2.0
h11==0.16.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
locust==2.43.0
lupa==2.6
MarkupSafe==3.0.3
model-bakery==1.21.0
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.9
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.13.0
python-json-logger==2.0.7
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==4.6.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simple-websocket==1.1.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uuid6==2025.0.1
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
